aider: Add score and time duration display in the top right corner of the screen.
# Aider chat conversation:

USER: now add a score of how many pieces of food the snake has eaten in the top right with large font. Also add a time duration in milliseconds.

ASSISTANT: To add a score and a timer, we need to modify the `snake.py` file. We will use the `curses` library's functionality to print text on the screen.

Here are the changes:

```python
snake.py
<<<<<<< ORIGINAL
score = 0

key = curses.KEY_RIGHT

while True:
    next_key = w.getch()
    key = key if next_key == -1 else next_key

    if snake[0][0] in [0, sh] or \
        snake[0][1]  in [0, sw] or \
        snake[0] in snake[1:]:
        curses.endwin()
        quit()
=======
import time

score = 0
start_time = time.time()

key = curses.KEY_RIGHT

while True:
    elapsed_time = time.time() - start_time
    w.addstr(0, sw-12, 'Score : ' + str(score) + ' Time : ' + str(int(elapsed_time*1000)) + ' ms')

    next_key = w.getch()
    key = key if next_key == -1 else next_key

    if snake[0][0] in [0, sh] or \
        snake[0][1]  in [0, sw] or \
        snake[0] in snake[1:]:
        curses.endwin()
        quit()
>>>>>>> UPDATED
```

In the updated code, we first import the `time` module. We then record the start time before the game loop starts. Inside the game loop, we calculate the elapsed time since the start of the game, convert it to milliseconds, and display it along with the score at the top right of the screen.